requirements.txt
# PostgreSQL driver
psycopg2-binary==2.9.9

# Data Processing
pyspark==3.5.1

# MinIO / S3-compatible object storage
minio==7.2.7
s3fs==2024.6.1
boto3==1.34.144

python-dotenv==1.0.1